+++ filter/all.go
@@ -11,6 +11,7 @@ var All = []baker.FilterDesc{
 	ConcatenateDesc,
 	MatchRegexDesc,
 	NotNullDesc,
+	ReplaceFieldsDesc,
 	SetStringFromURLDesc,
 	TimestampRangeDesc,
 }